refactor: extract complexity from StorageGrid
Extracted the deeply nested rendering logic in `StorageGrid.tsx` into smaller, reusable components: `StorageLocationSection`, `StorageCard`, and `EmptyLocation`. This improves code readability and maintainability.
